unsigned int -> uint16_t (oops)

+++ include/soul/renderer.hh
@@ -37,7 +37,7 @@ class Renderer {
 	public:
 		~Renderer();
 		static tl::expected<Renderer*, Error> create(Window* window);
-		void setVertexAndIndexBuffers(std::vector<Vertex>& verts, std::vector<unsigned int> indices);
+		void setVertexAndIndexBuffers(std::vector<Vertex>& verts, std::vector<uint16_t> indices);
 		Error update();
 	private:
 		Renderer(Window* new_window, bgfx::ProgramHandle new_program, bgfx::DynamicVertexBufferHandle new_vertex_buffer, bgfx::DynamicIndexBufferHandle new_index_buffer);
